# openapi.yaml
openapi: 3.0.3
info:
  title: Acme Orders API
  version: 1.0.0
  description: |
    API for managing customers, products, and orders for Acme Inc.
    Includes authentication, nested models, and realistic data structures.
servers:
  - url: http://localhost:8000/api/v1

tags:
  - name: Customers
    description: Customer management
  - name: Products
    description: Product management
  - name: Orders
    description: Order management

paths:
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      tags: [Customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Created customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

  /customers/{customer_id}:
    get:
      tags: [Customers]
      summary: Get customer by ID
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found

  /products:
    get:
      tags: [Products]
      summary: List products
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

  /orders:
    post:
      tags: [Orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

components:
  schemas:
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        status:
          $ref: '#/components/schemas/CustomerStatus'
    CustomerCreate:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
    CustomerStatus:
      type: string
      enum: [active, inactive, banned]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
          format: float
    OrderCreate:
      type: object
      required: [customer_id, items]
      properties:
        customer_id:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderItem:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: integer
        quantity:
          type: integer

  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 10

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

security:
  - ApiKeyAuth: []